requirements.txt
requests
telebot
yt-dlp
mega.py
configparser